# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Zoom Notifier API
  description: Multi-tenant webhook relay and notification service for Zoom meetings.
  version: "2.0.0"
servers:
  - url: http://localhost:8888
    description: Local development
security: []
paths:
...
  /webhook/zoom:
    post:
      operationId: postWebhookZoom
      summary: Receive Zoom webhook events
      tags: [zoom]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
...
  /api/v1/tenants/{tenantId}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      operationId: getTenant
      summary: Get tenant details
      tags: [tenants]
      security:
        - TenantKey: []
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTenant
      summary: Delete a tenant
      tags: [tenants]
      security:
        - TenantKey: []
      responses:
        "204":
          description: Tenant deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    AdminKey:
      type: http
      scheme: bearer
      description: Deployment-level admin API key for tenant management
    TenantKey:
      type: http
      scheme: bearer
      description: Per-tenant API key for tenant-scoped operations
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Tenant:
      type: object
      required: [id, api_key]
      properties:
        id:
          type: string
        team_name:
          type: string
        api_key:
          type: string
        installed_at:
          type: string
          format: date-time
        zoom_account_id:
          type: string
        default_msg_suffix:
          type: string
        default_include_link:
          type: boolean
...